backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
pillow>=10.3.0
qrcode[pil]>=7.4.2
openpyxl>=3.1.2
numpy>=1.26.0
matplotlib>=3.8.0
aiofiles>=23.2.1
